moves all cc and h files to src/
