+++ src/constants.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "constants.h"
+#include "src/constants.h"
 
 #include "glog/logging.h"
