+++ src/fido2_conformance_main.cc
@@ -14,13 +14,13 @@
 
 #include <iostream>
 
-#include "constants.h"
-#include "device_tracker.h"
 #include "gflags/gflags.h"
 #include "glog/logging.h"
-#include "hid/hid_device.h"
-#include "parameter_check.h"
-#include "test_series.h"
+#include "src/constants.h"
+#include "src/device_tracker.h"
+#include "src/hid/hid_device.h"
+#include "src/parameter_check.h"
+#include "src/test_series.h"
 
 DEFINE_string(
     token_path, "",
