+++ src/parameter_check.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "parameter_check.h"
+#include "src/parameter_check.h"
 
 #include <algorithm>
